refactor(utils): rename @libs/utils -> @internals/utils
