+++ .changeset/brown-cats-buy.md
@@ -1,0 +1,5 @@
+---
+'@internals/utils': minor
+---
+
+rename `@libs/utils` to `@internals/utils`
