+++ .github/workflows/deploy_www.yaml
@@ -40,7 +40,7 @@ jobs:
           paths_filter: |
             www:
               paths:
-                - 'libs/utils/**'
+                - 'internals/utils/**'
                 - 'sites/www/**'
   www:
     needs: pre_jobs
