+++ .github/workflows/unit_tests.yaml
@@ -5,7 +5,7 @@ on:
   pull_request:
     branches: [main]
     paths:
-      - 'libs/utils/**'
+      - 'internals/utils/**'
       - '**/vitest.config.js'
       - '.github/workflows/unit_tests.yaml'
 
@@ -23,7 +23,7 @@ jobs:
         with:
           filters: |
             utils:
-              - 'libs/utils/**.js'
+              - 'internals/utils/**.js'
               - '.github/workflows/unit_tests.yaml'
 
   setup:
@@ -102,14 +102,14 @@ jobs:
 
         # TODO: improve with `vitest latest` for only changed files
       - name: Run unit tests
-        run: pnpm test:ci --filter=@libs/utils
+        run: pnpm test:ci --filter=@internals/utils
         continue-on-error: false
 
         # TODO: simplify report steps with https://github.com/davelosert/vitest-coverage-report-action
       - name: Code coverage
         uses: irongut/CodeCoverageSummary@v1.3.0
         with:
-          filename: libs/utils/coverage/cobertura-coverage.xml
+          filename: internals/utils/coverage/cobertura-coverage.xml
           badge: true
           fail_below_min: true
           format: markdown
@@ -121,7 +121,7 @@ jobs:
         uses: marocchino/sticky-pull-request-comment@v2
         if: github.event_name == 'pull_request'
         with:
-          header: '@libs/utils'
+          header: '@internals/utils'
           hide_and_recreate: true
           hide_classify: "OUTDATED"
           path: code-coverage-results.md
