+++ internals/utils/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@libs/utils",
+  "name": "@internals/utils",
   "author": {
     "email": "developers@sveltevietnam.dev",
     "name": "Svelte Vietnam",
