+++ internals/utils/src/signature/index.js
