+++ internals/utils/src/signature/signature/signature.js
