+++ internals/utils/src/url/delocalize-url/delocalize-url.js
