+++ internals/utils/src/url/is-url-localized/is-url-localized.js
