+++ pnpm-workspace.yaml
@@ -1,4 +1,4 @@
 packages:
   - "services/*"
   - "sites/*"
-  - "libs/*"
+  - "internals/*"
