+++ sites/www/package.json
@@ -34,7 +34,7 @@
     "@builder.io/partytown": "^0.8.1",
     "@cloudflare/workers-types": "^4.20230904.0",
     "@csstools/postcss-global-data": "^2.1.0",
-    "@libs/utils": "workspace:^",
+    "@internals/utils": "workspace:^",
     "@svelte-put/clickoutside": "^3.0.0",
     "@svelte-put/modal": "^1.0.2",
     "@svelte-put/noti": "^1.1.0",
